+++ project/pyspark/{{ cookiecutter.project_name }}/dags/{{ cookiecutter.project_name }}.py
@@ -32,15 +32,7 @@ sample_task = ConveyorSparkSubmitOperatorV2(
     {%- if cookiecutter.conveyor_managed_role %}
     aws_role="{{ cookiecutter.project_name }}-{% raw %}{{ macros.conveyor.env() }}{% endraw %}",
     {%- endif %}
-    {% if cookiecutter.spark_version == "2.4" -%}
-    spark_main_version=2,
-    {%- elif cookiecutter.spark_version == "3.0" -%}
     spark_main_version=3,
-    {%- endif %}
-    {% if cookiecutter.spark_version == "2.4" -%}
-    application="/opt/spark/work-dir/src/{{ cookiecutter.module_name }}/app.py",
-    {%- elif cookiecutter.spark_version == "3.0" -%}
     application="local:///opt/spark/work-dir/src/{{ cookiecutter.module_name }}/app.py",
-    {%- endif %}
     application_args=["{% raw %}--date", "{{ ds }}", "--env", "{{ macros.conveyor.env() }}{% endraw %}"],
 )